+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2020-07-12
 ### Changed
-- ``IsDef.return_type(::Type{<:Tuple})`` now takes a single Tuple type, which includes the function type
+- `IsDef.return_type(::Type{<:Tuple})` now takes a single Tuple type, which includes the function type
 ### Added
 - when you overload `IsDef.return_type` for your custom type inference fixes, this is now also taken into account when using `apply`
 - GithubActions for CI, Codecov and Documentation
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.0] - 2020-03-23
 ### Changed
-- renamed ``IsDef._return_type`` to ``IsDef.return_type`` for analogy with `Core.Compiler.return_type`
+- renamed `IsDef._return_type` to `IsDef.return_type` for analogy with `Core.Compiler.return_type`
 - `Function` is now dealt like `Any` in the sense that no attempt is made to break it down into subtypes
 
 ### Fixed